Allow configuration of certain GPMC parameters
Allow the configuration of the M_NAND_GPMC and STNOR_GPMC config
register values as well as the GPMC config parameters for pin polarity
etc.

If not defined in the board's config, default values will be used as
they were defined before. This makes the GPMC code slightly more
generic, which is useful for custom boards.

Signed-off-by: Ronald Landheer-Cieslak <ronaldlandheercieslak@eaton.com>

%% original patch: 0014-Allow-configuration-of-certain-GPMC-parameters.patch