Implement passing boot source from SPL to boot
This is a feature we used in the previous version of the boot, but which
was not documented and broke when updating.

Basically, the SPL needs to tell the boot which device we're booting
from to make sure we don't ignore the jumper on the carrier board. So,
if the Sitara decided to boot from MMC we should continue booting from
MMC and if the Sitara decided to boot from SPI we should continue
booting from SPI.

The way this works now is by having the SPL pass the boot the source
device as its argument when jumping into the boot (so it's in r0). As
soon as the boot starts, it copies the value to r9, which is defined as
platform-specific data. r9 will eventually contain a pointer to the
global data structure, but all of that isn't set up yet at that point.

When the CPU is initialized and memory is set up, the global data is
also set up. At that time, the value is recovered (it's set aside before
relocating, recovered while relocating, set aside again, loaded into r9
an finally put back in global data, all to conserve it while the boot is
doing its thing) and put in its final location in the global data
structure.

I haven't found any easier way to do this, that would require less
juggling with the value.

* Note that the save_boot_params function is no longer used in either the
  SPL or the boot: the start.S file does the job in both cases.
* Also note that the SPL used to pass a pointer to the global data,
  which was not used. it now passes the only value we're interested in.
* note I've resurrected the arch_setup_gd function, which already exists
  on other platforms. This code won't compile on other ARM chips unless
  a version of that function is added to those platforms -- a version
  that would probably not need to do anything.